requirements.txt
Django==4.0.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0
django-extensions==3.2.0
ipython==8.4.0
mysqlclient==2.1.1
PyMySQL==1.0.2
python-dotenv==0.20.0
bcrypt==3.2.0
drf-yasg==1.20.3
cffi==1.15.1